+++ binance_testnet_agent/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "1.0.18"
+__version__ = "1.0.19"